pyproject.toml
[tool.poetry]
name = "test_cam"
version = "0.1.0"
description = ""
authors = ["xh4vm <h4vm@yandex.ru>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-split-settings = "^1.2.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
django-debug-toolbar = "^4.2.0"
django-cors-headers = "^4.2.0"
jsonschema = "^4.19.0"
loguru = "^0.7.1"
pydantic = "^2.3.0"
faker = "^19.6.0"
aiohttp = "^3.8.5"
pydantic-settings = "^2.0.3"
asyncclick = "^8.1.3.4"
anyio = "^4.0.0"
drf-yasg = "^1.21.7"
black = "^23.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"